+++ src/services/TranslateService.php
@@ -14,8 +14,7 @@ use craft\elements\Entry;
 use craft\fields\Table;
 use craft\models\Section;
 use craft\models\Site;
-use digitalpulsebe\craftmultitranslator\helpers\EntryHelper;
-use digitalpulsebe\craftmultitranslator\helpers\ProductHelper;
+use digitalpulsebe\craftmultitranslator\helpers\ElementHelper;
 use digitalpulsebe\craftmultitranslator\MultiTranslator;
 
 class TranslateService extends Component
@@ -24,6 +23,7 @@ class TranslateService extends Component
         'craft\fields\PlainText',
         'craft\redactor\Field',
         'craft\ckeditor\Field',
+        'abmat\tinymce\Field',
     ];
     static array $matrixFields = [
         'craft\fields\Matrix',
@@ -151,7 +151,10 @@ class TranslateService extends Component
                 $translatedValue = $this->translateVizyField($source, $field, $sourceSite, $targetSite);
             }
 
-            if (get_class($field) == 'craft\ckeditor\Field') {
+            if (in_array(get_class($field), [
+                'craft\ckeditor\Field',
+                'abmat\tinymce\Field',
+            ])) {
                 // search for interal href links
                 $translatedValue = $this->translateLinks($translatedValue, $sourceSite, $targetSite);
             }
@@ -323,7 +326,9 @@ class TranslateService extends Component
     public function findTargetElement(Element $source, int $targetSiteId): Element
     {
         if ($source instanceof Product) {
-            return ProductHelper::one($source->id, $targetSiteId);
+            return ElementHelper::one(Product::class, $source->id, $targetSiteId);
+        } elseif ($source instanceof Asset) {
+            return ElementHelper::one(Asset::class, $source->id, $targetSiteId);
         } elseif ($source instanceof Variant) {
             return Variant::find()->status(null)->id($source->id)->siteId($targetSiteId)->one();
         } else {
@@ -333,7 +338,7 @@ class TranslateService extends Component
 
     public function findTargetEntry(Entry $source, int $targetSiteId): Entry
     {
-        $targetEntry = EntryHelper::one($source->id, $targetSiteId);
+        $targetEntry = ElementHelper::one(Entry::class, $source->id, $targetSiteId);
 
         if (empty($targetEntry)) {
             // we need to create one for this target site
@@ -357,7 +362,7 @@ class TranslateService extends Component
                     Craft::$app->elements->saveElement($source);
                 }
 
-                $targetEntry = EntryHelper::one($source->id, $targetSiteId);
+                $targetEntry = ElementHelper::one(Entry::class, $source->id, $targetSiteId);
             } elseif ($source->section->propagationMethod == Section::PROPAGATION_METHOD_ALL) {
                 // it should have been there, so propagate
                 $targetEntry = Craft::$app->elements->propagateElement($source, $targetSiteId, false);