+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Multi Translator
 
+## 1.11.0 - 2024-07-09
+
+### Added
+- added support for Asset Elements
+- added support for abmat\tinymce\Field; PR from @HigumaSan4050
+
 ## 1.10.2 - 2024-06-18
 
 ### Fixed